graph_rnaseq_counts.R: Change shape option to shapeVariable.
The shapeVariable option now maps shape to a column in the samples file
and not just condition. 'none' is the default and in this case
shape isn't used.
shapeVariable should be one of the columns in the samples file.
colourVariable should also be one of the columns in the samples file.